requirements_stic.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.4.0
async-timeout==4.0.3
attrs==23.2.0
av==12.1.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
cachetools==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.5.1
contourpy==1.2.1
cycler==0.12.1
datasets==2.16.1
decord==0.6.0
deepspeed==0.12.2
dill==0.3.7
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
email_validator==2.1.2
exceptiongroup==1.2.1
fairscale==0.4.13
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.3
flash-attn==2.6.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gradio==3.35.2
gradio_client==0.2.9
grpcio==1.65.0
h11==0.14.0
hf_transfer==0.1.8
hjson==3.1.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.7
-e git+https://github.com/jena-shreyas/STIC@a033684e298a21d88843f32f282fda53a7591932#egg=llamavid&subdirectory=LLaMA-VID
-e git+https://github.com/jena-shreyas/STIC@a033684e298a21d88843f32f282fda53a7591932#egg=llava&subdirectory=LLaVA-NeXT
Markdown==3.6
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.9.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
networkx==3.3
ninja==1.11.1.1
ntfy==2.7.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
open_clip_torch==2.26.1
openai==0.28.0
opencv-python==4.10.0.84
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
peft==0.4.0
pillow==10.3.0
pipx==1.6.0
platformdirs==4.2.2
portalocker==2.8.2
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==1.10.8
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge==1.0.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.4.9
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.6.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
slacker==0.14.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==0.6.13
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torch-tb-profiler==0.4.3
torchvision==0.16.2
tqdm==4.66.4
transformers==4.42.0
triton==2.1.0
trl==0.9.3
typer==0.12.3
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==1.26.19
userpath==1.9.2
uvicorn==0.30.1
uvloop==0.19.0
-e git+https://github.com/jena-shreyas/STIC.git@536d8a95eb5f507d7bdc0b8a5f849c968a5b83e6#egg=videollava&subdirectory=VideoLLaVA
wandb==0.16.5
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
